Remove broken version of getMappingIterator
The transformation function f is of type Inner -> ResultType,
but Inner is the type of the inner iterator and not the type
of elements.  So this version always leads to a substitution failure.

Since the other overload should cover the intended use case anyway,
we can just remove the broken version.